+++ tests/Grant/DeviceCodeGrantTest.php
@@ -20,7 +20,6 @@ use League\OAuth2\Server\Repositories\RefreshTokenRepositoryInterface;
 use League\OAuth2\Server\Repositories\ScopeRepositoryInterface;
 use LeagueTests\Stubs\AccessTokenEntity;
 use LeagueTests\Stubs\ClientEntity;
-use LeagueTests\Stubs\CryptTraitStub;
 use LeagueTests\Stubs\DeviceCodeEntity;
 use LeagueTests\Stubs\RefreshTokenEntity;
 use LeagueTests\Stubs\ScopeEntity;
@@ -28,22 +27,13 @@ use LeagueTests\Stubs\StubResponseType;
 use PHPUnit\Framework\TestCase;
 
 use function base64_encode;
-use function json_encode;
 use function random_bytes;
-use function time;
 use function uniqid;
 
 class DeviceCodeGrantTest extends TestCase
 {
     private const DEFAULT_SCOPE = 'basic';
     private const INTERVAL_RATE = 10;
-
-    protected CryptTraitStub $cryptStub;
-
-    public function setUp(): void
-    {
-        $this->cryptStub = new CryptTraitStub();
-    }
 
     public function testGetIdentifier(): void
     {
@@ -102,7 +92,6 @@ class DeviceCodeGrantTest extends TestCase
 
         $grant->setClientRepository($clientRepositoryMock);
         $grant->setDefaultScope(self::DEFAULT_SCOPE);
-        $grant->setEncryptionKey($this->cryptStub->getKey());
         $grant->setScopeRepository($scopeRepositoryMock);
 
         $request = (new ServerRequest())->withParsedBody([
@@ -147,7 +136,6 @@ class DeviceCodeGrantTest extends TestCase
 
         $grant->setClientRepository($clientRepositoryMock);
         $grant->setDefaultScope(self::DEFAULT_SCOPE);
-        $grant->setEncryptionKey($this->cryptStub->getKey());
         $grant->setScopeRepository($scopeRepositoryMock);
 
         $request = (new ServerRequest())->withParsedBody([
@@ -274,8 +262,6 @@ class DeviceCodeGrantTest extends TestCase
             'http://foo/bar',
         );
 
-        $grant->setEncryptionKey($this->cryptStub->getKey());
-
         $grant->completeDeviceAuthorizationRequest($deviceCode->getIdentifier(), 'userId', true);
 
         $this::assertEquals('userId', $deviceCode->getUserIdentifier());
@@ -325,8 +311,6 @@ class DeviceCodeGrantTest extends TestCase
             'http://foo/bar'
         );
 
-        $deviceCodeGrant->setEncryptionKey($this->cryptStub->getKey());
-
         $server->enableGrantType($deviceCodeGrant);
 
         $response = $server->respondToDeviceAuthorizationRequest($serverRequest, new Response());
@@ -350,28 +334,35 @@ class DeviceCodeGrantTest extends TestCase
         $scope = new ScopeEntity();
         $scope->setIdentifier('foo');
 
+        $deviceCodeRepositoryMock = $this->getMockBuilder(DeviceCodeRepositoryInterface::class)->getMock();
+        $deviceCodeEntity = new DeviceCodeEntity();
+
+        $deviceCodeEntity->setUserIdentifier('baz');
+        $deviceCodeEntity->setIdentifier('deviceCodeEntityIdentifier');
+        $deviceCodeEntity->setUserCode('123456');
+        $deviceCodeEntity->setExpiryDateTime(new DateTimeImmutable('+1 hour'));
+        $deviceCodeEntity->setClient($client);
+        $deviceCodeEntity->addScope($scope);
+
+        $deviceCodeRepositoryMock->method('getDeviceCodeEntityByDeviceCode')
+            ->with($deviceCodeEntity->getIdentifier())
+            ->willReturn($deviceCodeEntity);
+
         $accessTokenEntity = new AccessTokenEntity();
         $accessTokenEntity->addScope($scope);
+
         $accessTokenRepositoryMock = $this->getMockBuilder(AccessTokenRepositoryInterface::class)->getMock();
-        $accessTokenRepositoryMock->method('getNewToken')->willReturn($accessTokenEntity);
+        $accessTokenRepositoryMock->method('getNewToken')
+            ->with($client, $deviceCodeEntity->getScopes(), $deviceCodeEntity->getUserIdentifier())
+            ->willReturn($accessTokenEntity);
         $accessTokenRepositoryMock->method('persistNewAccessToken')->willReturnSelf();
 
         $refreshTokenRepositoryMock = $this->getMockBuilder(RefreshTokenRepositoryInterface::class)->getMock();
         $refreshTokenRepositoryMock->method('persistNewRefreshToken')->willReturnSelf();
         $refreshTokenRepositoryMock->method('getNewRefreshToken')->willReturn(new RefreshTokenEntity());
 
-        $deviceCodeRepositoryMock = $this->getMockBuilder(DeviceCodeRepositoryInterface::class)->getMock();
-        $deviceCodeEntity = new DeviceCodeEntity();
-        $deviceCodeEntity->setUserIdentifier('baz');
-        $deviceCodeEntity->setIdentifier('deviceCodeEntityIdentifier');
-        $deviceCodeEntity->setUserCode('123456');
-        $deviceCodeEntity->setExpiryDateTime(new DateTimeImmutable('+1 hour'));
-        $deviceCodeEntity->setClient($client);
-
-        $deviceCodeRepositoryMock->method('getDeviceCodeEntityByDeviceCode')->willReturn($deviceCodeEntity);
-
         $scopeRepositoryMock = $this->getMockBuilder(ScopeRepositoryInterface::class)->getMock();
-        $scopeRepositoryMock->method('getScopeEntityByIdentifier')->willReturn($scope);
+        $scopeRepositoryMock->expects(self::never())->method('getScopeEntityByIdentifier');
         $scopeRepositoryMock->method('finalizeScopes')->willReturnArgument(0);
 
         $grant = new DeviceCodeGrant(
@@ -385,24 +376,13 @@ class DeviceCodeGrantTest extends TestCase
         $grant->setAccessTokenRepository($accessTokenRepositoryMock);
         $grant->setScopeRepository($scopeRepositoryMock);
         $grant->setDefaultScope(self::DEFAULT_SCOPE);
-        $grant->setEncryptionKey($this->cryptStub->getKey());
         $grant->setPrivateKey(new CryptKey('file://' . __DIR__ . '/../Stubs/private.key'));
 
         $grant->completeDeviceAuthorizationRequest($deviceCodeEntity->getIdentifier(), 'baz', true);
 
         $serverRequest = (new ServerRequest())->withParsedBody([
             'grant_type' => 'urn:ietf:params:oauth:grant-type:device_code',
-            'device_code'   => $this->cryptStub->doEncrypt(
-                json_encode(
-                    [
-                        'device_code_id' => 'deviceCodeEntityIdentifier',
-                        'expire_time' => time() + 3600,
-                        'client_id' => 'foo',
-                        'scopes' => ['foo'],
-                    ],
-                    JSON_THROW_ON_ERROR
-                )
-            ),
+            'device_code'   => $deviceCodeEntity->getIdentifier(),
             'client_id'     => 'foo',
         ]);
 
@@ -433,17 +413,7 @@ class DeviceCodeGrantTest extends TestCase
         $grant->setAccessTokenRepository($accessTokenRepositoryMock);
 
         $serverRequest = (new ServerRequest())->withQueryParams([
-            'device_code' => $this->cryptStub->doEncrypt(
-                json_encode(
-                    [
-                        'device_code_id' => uniqid(),
-                        'expire_time' => time() + 3600,
-                        'client_id' => 'foo',
-                        'scopes' => ['foo'],
-                    ],
-                    JSON_THROW_ON_ERROR
-                )
-            ),
+            'device_code' => uniqid(),
         ]);
 
         $responseType = new StubResponseType();
@@ -470,9 +440,8 @@ class DeviceCodeGrantTest extends TestCase
         $deviceCodeEntity->setUserIdentifier('baz');
         $deviceCodeRepositoryMock->method('getDeviceCodeEntityByDeviceCode')->willReturn($deviceCodeEntity);
 
-        $scope = new ScopeEntity();
         $scopeRepositoryMock = $this->getMockBuilder(ScopeRepositoryInterface::class)->getMock();
-        $scopeRepositoryMock->method('getScopeEntityByIdentifier')->willReturn($scope);
+        $scopeRepositoryMock->expects(self::never())->method('getScopeEntityByIdentifier');
         $scopeRepositoryMock->method('finalizeScopes')->willReturnArgument(0);
 
         $grant = new DeviceCodeGrant(
@@ -485,7 +454,6 @@ class DeviceCodeGrantTest extends TestCase
         $grant->setClientRepository($clientRepositoryMock);
         $grant->setScopeRepository($scopeRepositoryMock);
         $grant->setDefaultScope(self::DEFAULT_SCOPE);
-        $grant->setEncryptionKey($this->cryptStub->getKey());
         $grant->setPrivateKey(new CryptKey('file://' . __DIR__ . '/../Stubs/private.key'));
 
         $serverRequest = (new ServerRequest())->withParsedBody([
@@ -519,10 +487,8 @@ class DeviceCodeGrantTest extends TestCase
         $deviceCodeEntity->setClient($client);
         $deviceCodeRepositoryMock->method('getDeviceCodeEntityByDeviceCode')->willReturn($deviceCodeEntity);
 
-        $scope = new ScopeEntity();
-        $scope->setIdentifier('foo');
         $scopeRepositoryMock = $this->getMockBuilder(ScopeRepositoryInterface::class)->getMock();
-        $scopeRepositoryMock->method('getScopeEntityByIdentifier')->willReturn($scope);
+        $scopeRepositoryMock->expects(self::never())->method('getScopeEntityByIdentifier');
         $scopeRepositoryMock->method('finalizeScopes')->willReturnArgument(0);
 
         $grant = new DeviceCodeGrant(
@@ -535,22 +501,11 @@ class DeviceCodeGrantTest extends TestCase
         $grant->setClientRepository($clientRepositoryMock);
         $grant->setScopeRepository($scopeRepositoryMock);
         $grant->setDefaultScope(self::DEFAULT_SCOPE);
-        $grant->setEncryptionKey($this->cryptStub->getKey());
         $grant->setPrivateKey(new CryptKey('file://' . __DIR__ . '/../Stubs/private.key'));
 
         $serverRequest = (new ServerRequest())->withParsedBody([
             'client_id'     => 'foo',
-            'device_code'   => $this->cryptStub->doEncrypt(
-                json_encode(
-                    [
-                        'device_code_id' => uniqid(),
-                        'expire_time' => time() + 3600,
-                        'client_id' => 'foo',
-                        'scopes' => ['foo'],
-                    ],
-                    JSON_THROW_ON_ERROR
-                )
-            ),
+            'device_code'   => uniqid(),
         ]);
 
         $responseType = new StubResponseType();
@@ -579,10 +534,8 @@ class DeviceCodeGrantTest extends TestCase
         $deviceCodeEntity->setClient($client);
         $deviceCodeRepositoryMock->method('getDeviceCodeEntityByDeviceCode')->willReturn($deviceCodeEntity);
 
-        $scope = new ScopeEntity();
-        $scope->setIdentifier('foo');
         $scopeRepositoryMock = $this->getMockBuilder(ScopeRepositoryInterface::class)->getMock();
-        $scopeRepositoryMock->method('getScopeEntityByIdentifier')->willReturn($scope);
+        $scopeRepositoryMock->expects(self::never())->method('getScopeEntityByIdentifier');
         $scopeRepositoryMock->method('finalizeScopes')->willReturnArgument(0);
 
         $grant = new DeviceCodeGrant(
@@ -595,22 +548,11 @@ class DeviceCodeGrantTest extends TestCase
         $grant->setClientRepository($clientRepositoryMock);
         $grant->setScopeRepository($scopeRepositoryMock);
         $grant->setDefaultScope(self::DEFAULT_SCOPE);
-        $grant->setEncryptionKey($this->cryptStub->getKey());
         $grant->setPrivateKey(new CryptKey('file://' . __DIR__ . '/../Stubs/private.key'));
 
         $serverRequest = (new ServerRequest())->withParsedBody([
             'client_id'     => 'foo',
-            'device_code'   => $this->cryptStub->doEncrypt(
-                json_encode(
-                    [
-                        'device_code_id' => uniqid(),
-                        'expire_time' => time() + 3600,
-                        'client_id' => 'foo',
-                        'scopes' => ['foo'],
-                    ],
-                    JSON_THROW_ON_ERROR
-                )
-            ),
+            'device_code'   => uniqid(),
         ]);
 
         $responseType = new StubResponseType();
@@ -639,9 +581,8 @@ class DeviceCodeGrantTest extends TestCase
         $deviceCodeEntity->setClient($client);
         $deviceCodeRepositoryMock->method('getDeviceCodeEntityByDeviceCode')->willReturn($deviceCodeEntity);
 
-        $scope = new ScopeEntity();
         $scopeRepositoryMock = $this->getMockBuilder(ScopeRepositoryInterface::class)->getMock();
-        $scopeRepositoryMock->method('getScopeEntityByIdentifier')->willReturn($scope);
+        $scopeRepositoryMock->expects(self::never())->method('getScopeEntityByIdentifier');
         $scopeRepositoryMock->method('finalizeScopes')->willReturnArgument(0);
 
         $grant = new DeviceCodeGrant(
@@ -654,22 +595,11 @@ class DeviceCodeGrantTest extends TestCase
         $grant->setClientRepository($clientRepositoryMock);
         $grant->setScopeRepository($scopeRepositoryMock);
         $grant->setDefaultScope(self::DEFAULT_SCOPE);
-        $grant->setEncryptionKey($this->cryptStub->getKey());
         $grant->setPrivateKey(new CryptKey('file://' . __DIR__ . '/../Stubs/private.key'));
 
         $serverRequest = (new ServerRequest())->withParsedBody([
             'client_id'     => 'foo',
-            'device_code'   => $this->cryptStub->doEncrypt(
-                json_encode(
-                    [
-                        'device_code_id' => uniqid(),
-                        'expire_time' => time() - 3600,
-                        'client_id' => 'foo',
-                        'scopes' => ['foo'],
-                    ],
-                    JSON_THROW_ON_ERROR
-                )
-            ),
+            'device_code'   => uniqid(),
         ]);
 
         $responseType = new StubResponseType();
@@ -708,7 +638,6 @@ class DeviceCodeGrantTest extends TestCase
 
         $grant->setClientRepository($clientRepositoryMock);
         $grant->setDefaultScope(self::DEFAULT_SCOPE);
-        $grant->setEncryptionKey($this->cryptStub->getKey());
         $grant->setScopeRepository($scopeRepositoryMock);
         $grant->setIntervalVisibility(true);
 
@@ -749,10 +678,8 @@ class DeviceCodeGrantTest extends TestCase
         $deviceCode->setUserCode('12345678');
         $deviceCodeRepositoryMock->method('getDeviceCodeEntityByDeviceCode')->willReturn($deviceCode);
 
-        $scope = new ScopeEntity();
-        $scope->setIdentifier('foo');
         $scopeRepositoryMock = $this->getMockBuilder(ScopeRepositoryInterface::class)->getMock();
-        $scopeRepositoryMock->method('getScopeEntityByIdentifier')->willReturn($scope);
+        $scopeRepositoryMock->expects(self::never())->method('getScopeEntityByIdentifier');
         $scopeRepositoryMock->method('finalizeScopes')->willReturnArgument(0);
 
         $grant = new DeviceCodeGrant(
@@ -765,24 +692,13 @@ class DeviceCodeGrantTest extends TestCase
         $grant->setClientRepository($clientRepositoryMock);
         $grant->setScopeRepository($scopeRepositoryMock);
         $grant->setDefaultScope(self::DEFAULT_SCOPE);
-        $grant->setEncryptionKey($this->cryptStub->getKey());
         $grant->setPrivateKey(new CryptKey('file://' . __DIR__ . '/../Stubs/private.key'));
 
         $grant->completeDeviceAuthorizationRequest($deviceCode->getIdentifier(), '1', false);
 
         $serverRequest = (new ServerRequest())->withParsedBody([
                 'client_id'     => 'foo',
-                'device_code'   => $this->cryptStub->doEncrypt(
-                    json_encode(
-                        [
-                        'device_code_id' => uniqid(),
-                        'expire_time' => time() + 3600,
-                        'client_id' => 'foo',
-                        'scopes' => ['foo'],
-                        ],
-                        JSON_THROW_ON_ERROR
-                    ),
-                ),
+                'device_code'   => $deviceCode->getIdentifier(),
         ]);
 
         $responseType = new StubResponseType();